+++ src/natcap/invest/routedem.py
@@ -13,7 +13,7 @@ from . import spec_utils
 from . import utils
 from . import validation
 from .model_metadata import MODEL_METADATA
-from .spec_utils import u
+from .unit_registry import u
 
 LOGGER = logging.getLogger(__name__)