+++ src/natcap/invest/seasonal_water_yield/seasonal_water_yield.py
@@ -18,7 +18,7 @@ from .. import spec_utils
 from .. import utils
 from .. import validation
 from ..model_metadata import MODEL_METADATA
-from ..spec_utils import u
+from ..unit_registry import u
 from . import seasonal_water_yield_core
 
 gdal.SetCacheMax(2**26)